Consensus sorting is implemented but does not work.
Whether the inability of consensus sorting to produce reasonable
output is a result of bugs or intrinsic to the idea is not yet
clear. More work will be done to sort this out.

As a new feature, grids with lower dimensionality than their
container can now be interpolated (for 3D at least). This is
achieved by forcing otherwise out-of-bounds interpolation points
to have an exact match for the singleton grid dimensions.
This should be useful, e.g., when a 1-D or 2-D system is to be
compared to measured data as the instrumental effects can still
be modeled properly without also needing to replicate the
system response up to 3-D.

+++ src/arrayvector.h
@@ -83,20 +83,13 @@ public:
         // so calculate the span along each dimension of that row-ordered array
         std::vector<size_t> spans(shape.size());
         spans[shape.size()-1]=1u;
-        for (int i=shape.size()-2; i>-1; --i) spans[i] = spans[i+1]*shape[i+1];
+        for (int i=static_cast<int>(shape.size())-2; i>-1; --i)
+          spans[i] = spans[i+1]*shape[i+1];
         // if the calculated spans and input strides are equivalent, we can
         // skip calculating indicies:
         bool roword = true;
         for (int i=0; i<strides.size(); ++i)
           roword &= strides[i]/sizeof(T) == spans[i];
-        // std::cout << "New ArrayVector constructor passed"
-        //           << (roword ? " " : " non " ) << "row-ordered data."
-        //           << std::endl;
-        // std::cout << "Since input spans = [";
-        // for (int q=0; q<strides.size(); ++q) std::cout << " " << std::to_string(strides[q]/sizeof(T));
-        // std::cout << " ]\n and output spans = [";
-        // for (int q=0; q<spans.size(); ++q) std::cout << " " << std::to_string(spans[q]);
-        // std::cout << " ]" << std::endl;
         if (roword){
           for (size_t i=0; i<this->N*this->M; ++i) this->data[i] = d[i];
         } else {
